requirements.txt
aiogram
openai
python-dotenv
flask